# openapi.yaml
openapi: 3.0.3
info:
  title: LeadRecovery API
  description: >
    Multi-tenant lead recovery and reactivation API. Every route except
    /health, /ready, and the provider webhooks requires
    `Authorization: Bearer <tenant api key>`; admin routes require
    `Authorization: Bearer <ADMIN_API_KEY>` instead. See README.md for the
    full narrative reference — this file is the machine-readable version of
    the same routes.
  version: "0.2.0"
servers:
  - url: http://localhost:3000
    description: Local demo server (npm run dev)
...
security: []
paths:
...
  /tenants/me/team:
    get:
      tags: [team]
      summary: List this tenant's team members
      description: Open to any authenticated actor (owner or member) — see "Team accounts".
      security: [{ tenantAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PublicTenantUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Suspended" }
    post:
      tags: [team]
      summary: Invite a new team member
      description: >
        Owner-only. Emails the invited address a "set your password" link
        (src/authEmail.ts — logged server-side instead when no platform
        SendGrid key is configured). `email` must not already be in use by
        this tenant's own login or by another team member (globally,
        across every tenant).
      security: [{ tenantAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                role: { type: string, enum: [owner, member], default: member }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicTenantUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/OwnerRoleRequired" }
        "409": { description: "email already in use" }
  /tenants/me/team/{id}:
    patch:
      tags: [team]
      summary: Change a team member's role
      description: Owner-only.
      security: [{ tenantAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [owner, member] }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicTenantUser" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/OwnerRoleRequired" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [team]
      summary: Remove a team member
      description: Owner-only.
      security: [{ tenantAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Removed }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/OwnerRoleRequired" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    tenantAuth:
      type: http
      scheme: bearer
      description: A tenant's own API key.
    adminAuth:
      type: http
      scheme: bearer
      description: ADMIN_API_KEY.
...
  responses:
    Unauthorized:
      description: Missing or invalid Authorization header
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Suspended:
      description: Tenant is suspended
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    OwnerRoleRequired:
      description: Authenticated as a team member with role "member" — only the tenant's own login or an "owner"-role team member can do this (see "Team accounts")
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    PublicTenantUser:
      type: object
      description: >
        A team member under a tenant (see "Team accounts") with all secrets
        stripped (no passwordHash, no loginKey, no reset-token hash).
      properties:
        id: { type: string }
        tenantId: { type: string }
        email: { type: string }
        role: { type: string, enum: [owner, member] }
        hasPassword: { type: boolean, description: "Whether this member has ever set a password yet (false right after being invited, before they use the emailed link)." }
        passwordResetExpiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
...